Github Actions: Upload build artifacts, to be used by modules
The artifacts tarball is compressed with Zstandard to minimize delays
when uploading and downloading the sizeable binary artifacts.

Extra care has been taken to avoid using $HOME environment variable as
installation directory.

This is because $HOME is in mangled form under bash in windows
("/c/Users/runneradmin"), and cmake stores it in this form when passed
to `configure -prefix`, and it does not get demangled later, in some
custom CMD scripts during build of submodules.

Finally, a lot of the steps that are needed in common between qt modules
have been moved to scripts in the qt5 repo.

Some of the common steps, like setting up cache and uploading artifacts,
are still duplicated in the various Github Actions YAML files across
repositories.  This is because because custom "composite" actions
currently can't invoke other actions, the "uses" clause is denied. It's
on the roadmap though, see https://github.com/actions/runner/issues/646

Change-Id: I5b9c962c712b402c84bee1f6783ebcc0b86d4968